test: split via pad cases by custom rule
